% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_nca.R
\name{plot.nca_analysis}
\alias{plot.nca_analysis}
\title{Plot NCA Results}
\usage{
\method{plot}{nca_analysis}(x, type = c("scatter", "effects"), ...)
}
\arguments{
\item{x}{An \code{nca_analysis} object from \code{assess_nca()}.}

\item{type}{One of \code{"scatter"} (ceiling line scatter plots) or
\code{"effects"} (bar plot of effect sizes).}

\item{...}{Additional arguments passed to the underlying plot function.}
}
\description{
Plot NCA Results
}
