% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_nca.R
\name{plot.nca_esse}
\alias{plot.nca_esse}
\title{Plot NCA-ESSE Results}
\usage{
\method{plot}{nca_esse}(x, type = c("sensitivity", "difference"), ...)
}
\arguments{
\item{x}{An \code{nca_esse} object from \code{assess_nca_esse()}.}

\item{type}{One of \code{"sensitivity"} (effect size vs threshold, Fig. 4
in Becker et al.) or \code{"difference"} (incremental empirical minus
benchmark difference, Fig. 6 in Becker et al.).}

\item{...}{Additional arguments passed to the underlying plot function.}
}
\description{
Plot NCA-ESSE Results
}
