% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_cvpat.R
\name{assess_cvpat_compare}
\alias{assess_cvpat_compare}
\title{SEMinR function to compare CV-PAT loss of two models}
\usage{
assess_cvpat_compare(
  base_model,
  alt_sm,
  testtype = "two.sided",
  BootSamp = 2000,
  seed = 123,
  technique = predict_DA,
  noFolds = NULL,
  reps = NULL,
  cores = NULL
)
}
\arguments{
\item{base_model}{The base model for CV-PAT comparison.}

\item{alt_sm}{The alternate structural model for CV-PAT comparison.}

\item{testtype}{Either "two.sided" (default) or "greater".}

\item{BootSamp}{The number of bootstrap subsamples to execute (defaults to 2000).}

\item{seed}{The seed for reproducibility (defaults to 123).}

\item{technique}{predict_EA or predict_DA (default).}

\item{noFolds}{Mumber of folds for k-fold cross validation.}

\item{reps}{Number of repetitions for cross validation.}

\item{cores}{Number of cores for parallelization.}
}
\value{
A matrix of the estimated loss and results of significance testing.
}
\description{
`assess_cvpat_compare` conducts a CV-PAT significance test of loss between
two models.
}
\examples{
# Load libraries
library(seminr)

# Create measurement model ----
corp_rep_mm_ext <- constructs(
 composite("QUAL", multi_items("qual_", 1:8), weights = mode_B),
 composite("PERF", multi_items("perf_", 1:5), weights = mode_B),
 composite("CSOR", multi_items("csor_", 1:5), weights = mode_B),
 composite("ATTR", multi_items("attr_", 1:3), weights = mode_B),
 composite("COMP", multi_items("comp_", 1:3)),
 composite("LIKE", multi_items("like_", 1:3))
)

# Create structural model ----

corp_rep_sm_ext <- relationships(
 paths(from = c("QUAL", "PERF", "CSOR", "ATTR"), to = c("COMP"))
)
alt_sm <- relationships(
 paths(from = c("QUAL", "PERF", "CSOR", "ATTR"), to = c("COMP", "LIKE"))
)

# Estimate the model ----
corp_rep_pls_model_ext <- estimate_pls(
 data = corp_rep_data,
 measurement_model = corp_rep_mm_ext,
 structural_model  = corp_rep_sm_ext,
 missing = mean_replacement,
 missing_value = "-99")

# Function to compare the Loss of two models
assess_cvpat_compare(base_model = corp_rep_pls_model_ext,
                    alt_sm = alt_sm,
                    testtype = "two.sided",
                    BootSamp = 100,
                    technique = predict_DA,
                    seed = 123,
                    cores = 1)

}
