% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_coa.R
\name{plot.coa_analysis}
\alias{plot.coa_analysis}
\title{Plot COA Results}
\usage{
\method{plot}{coa_analysis}(x, type = c("pd", "groups", "tree"), ...)
}
\arguments{
\item{x}{A `coa_analysis` object from `assess_coa()`.}

\item{type}{One of `"pd"` (predictive deviance scatter), `"groups"`
(average construct scores per group), or `"tree"` (rpart tree diagram).}

\item{...}{Additional arguments passed to the underlying plot function.}
}
\description{
Plot COA Results
}
